refactor: move TokenSampler tests to own class
